refactor(core): move LContainer manipulation logic to its own file
All the view in a container logic is in the
packages/core/src/render3/view/container.ts file
plus some other associated refactorings.
