+++ packages/core/test/bundling/defer/bundle.golden_symbols.json
@@ -432,6 +432,7 @@
   "init_constants",
   "init_construction",
   "init_container",
+  "init_container2",
   "init_context",
   "init_context_discovery",
   "init_contextual",
