+++ packages/core/test/bundling/hello_world/bundle.golden_symbols.json
@@ -157,6 +157,7 @@
   "consumerPollProducersForChange",
   "context",
   "convertToBitFlags",
+  "createDirectivesInstances",
   "createElementRef",
   "createErrorClass",
   "createInjector",
