+++ packages/core/test/bundling/hydration/bundle.golden_symbols.json
@@ -210,6 +210,7 @@
   "consumerPollProducersForChange",
   "context",
   "convertToBitFlags",
+  "createDirectivesInstances",
   "createElementNode",
   "createElementRef",
   "createErrorClass",
