+++ packages/core/test/bundling/standalone_bootstrap/bundle.golden_symbols.json
@@ -183,6 +183,7 @@
   "consumerPollProducersForChange",
   "context",
   "convertToBitFlags",
+  "createDirectivesInstances",
   "createElementRef",
   "createErrorClass",
   "createInjector",
