+++ packages/core/test/render3/di_spec.ts
@@ -7,7 +7,6 @@
  */
 
 import {Component, Directive, Self} from '@angular/core';
-import {createLView, createTView} from '@angular/core/src/render3/instructions/shared';
 import {NodeInjectorOffset} from '@angular/core/src/render3/interfaces/injector';
 import {TestBed} from '@angular/core/testing';
 
@@ -21,6 +20,7 @@ import {TNodeType} from '../../src/render3/interfaces/node';
 import {HEADER_OFFSET, LViewFlags, TVIEW, TViewType} from '../../src/render3/interfaces/view';
 import {enterView, leaveView} from '../../src/render3/state';
 import {getOrCreateTNode} from '@angular/core/src/render3/tnode_manipulation';
+import {createLView, createTView} from '@angular/core/src/render3/view/construction';
 
 describe('di', () => {
   describe('directive injection', () => {
