+++ packages/core/test/render3/instructions/shared_spec.ts
@@ -6,7 +6,6 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {createLView, createTView} from '@angular/core/src/render3/instructions/shared';
 import {TNodeType} from '@angular/core/src/render3/interfaces/node';
 import {
   HEADER_OFFSET,
@@ -24,6 +23,7 @@ import {
 
 import {MockRendererFactory} from './mock_renderer_factory';
 import {createTNode} from '@angular/core/src/render3/tnode_manipulation';
+import {createLView, createTView} from '@angular/core/src/render3/view/construction';
 
 /**
  * Setups a simple `LView` so that it is possible to do unit tests on instructions.
