+++ packages/core/test/render3/view_fixture.ts
@@ -12,7 +12,6 @@ import {stringifyElement} from '@angular/platform-browser/testing/src/browser_ut
 import {extractDirectiveDef} from '../../src/render3/definition';
 import {refreshView} from '../../src/render3/instructions/change_detection';
 import {renderView} from '../../src/render3/instructions/render';
-import {createLView, createTView} from '../../src/render3/instructions/shared';
 import {
   DirectiveDef,
   DirectiveDefList,
@@ -35,6 +34,7 @@ import {noop} from '../../src/util/noop';
 
 import {getRendererFactory2} from './imported_renderer2';
 import {createTNode} from '@angular/core/src/render3/tnode_manipulation';
+import {createLView, createTView} from '@angular/core/src/render3/view/construction';
 
 /**
  * Fixture useful for testing operations which need `LView` / `TView`
