+++ packages/core/test/render3/view_utils_spec.ts
@@ -6,10 +6,10 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {createLContainer} from '@angular/core/src/render3/instructions/shared';
 import {isLContainer, isLView} from '@angular/core/src/render3/interfaces/type_checks';
 import {ViewFixture} from './view_fixture';
 import {createTNode} from '@angular/core/src/render3/tnode_manipulation';
+import {createLContainer} from '@angular/core/src/render3/view/container';
 
 describe('view_utils', () => {
   it('should verify unwrap methods (isLView and isLContainer)', () => {
